+++ src/shared/util.cpp
@@ -81,12 +81,16 @@ std::wstring ToWString(const std::string &source, bool utf8)
     if (!utf8) {
       codepage = AreFileApisANSI() ? GetACP() : GetOEMCP();
     }
-    int sizeRequired = ::MultiByteToWideChar(codepage, 0, source.c_str(), source.length(), nullptr, 0);
+    int sizeRequired
+        = ::MultiByteToWideChar(codepage, 0, source.c_str(),
+                                static_cast<int>(source.length()), nullptr, 0);
     if (sizeRequired == 0) {
       throw windows_error("failed to convert string to wide character");
     }
     result.resize(sizeRequired, L'\0');
-    ::MultiByteToWideChar(codepage, 0, source.c_str(), source.length(), &result[0], sizeRequired);
+    ::MultiByteToWideChar(codepage, 0, source.c_str(),
+                          static_cast<int>(source.length()), &result[0],
+                          sizeRequired);
   }
 
   return result;
